nuevo todo

+++ routes/agenda.js
@@ -13,6 +13,7 @@ module.exports = function(app,ss) {
       nombre:  req.param('nombre'),
       insertado: fecha.getTime()
     });
+    agenda.centro='';
   
     if (isNaN(agenda.consulta) || !agenda.consulta) {
       switch (agenda.nombre)
@@ -37,7 +38,11 @@ module.exports = function(app,ss) {
                break;
                case 'EEM_ILLE': agenda.consulta='3';
                break;
-               default:  agenda.consulta=''
+               case 'EEM': agenda.consulta='3';
+               break;
+               case 'EEM2': agenda.consulta='3';
+               break;
+               default:  agenda.consulta='';
             };
       
     };
@@ -47,8 +52,41 @@ module.exports = function(app,ss) {
       agenda.consulta='ATS'
     }
 
-  //  agenda._id=agenda.nif + agenda.consulta;
-    ss.api.publish.all('newMessage', agenda);
+    //si contiene ILLE es ILLESCAS, sino parla
+    if(agenda.nombre.indexOf('_ILLE') > -1)
+    {
+      agenda.centro='ILLESCAS';
+    }
+    else
+    {
+      switch (agenda.nombre)
+            {
+               case 'MMR2': agenda.centro='PARLA2';
+               break; 
+               case 'EEM2': agenda.centro='PARLA2';
+               break;
+               default:  agenda.centro='PARLA';
+            };
+    }
+
+
+    //si contiene ILLE es ILLESCAS, sino parla
+    if(agenda.nombre.indexOf('_ILLE') > -1)
+    {
+      agenda.centro='ILLESCAS';
+      ss.api.publish.all('newMessage', agenda);
+    }
+    else if(agenda.centro.indexOf('PARLA2') > -1)
+    {
+      agenda.centro='PARLA2';
+      ss.api.publish.all('newMessageParla2', agenda);
+    }
+    else
+    {
+      agenda.centro='PARLA';
+      ss.api.publish.all('newMessageParla', agenda);
+    }
+  
     
 
     agenda.save(function(err) {
@@ -64,14 +102,29 @@ module.exports = function(app,ss) {
 
 
   rebootAgenda = function(req, res) {
-   
-    ss.api.publish.all('reboot', 'reboot');
+
+    var centro=req.param('CENTRO');
+
+    //si contiene ILLE es ILLESCAS, sino parla
+    if(centro.indexOf('ILLESCAS') > -1)
+    {
+      ss.api.publish.all('reboot', 'reboot');
+    }
+    else if(centro.indexOf('PARLA2') > -1)
+    {
+      ss.api.publish.all('rebootParla2', 'rebootParla2');
+    }
+    else
+    {
+      ss.api.publish.all('rebootParla', 'rebootParla');
+    }
 
     res.send(true);
   };
 
   app.post('/agenda', addAgenda);
   app.post('/reboot', rebootAgenda);
+
 
   app.get('/', function (req, res) {
         res.serve('main')}
